+++ braintree/configuration.py
@@ -1,9 +1,6 @@
 import os
 import sys
 import braintree
-import braintree.util.http_strategy.pycurl_strategy
-import braintree.util.http_strategy.httplib_strategy
-import braintree.util.http_strategy.requests_strategy
 
 class Configuration(object):
     """
@@ -16,47 +13,27 @@ class Configuration(object):
             "your_public_key",
             "your_private_key"
         )
-
-    By default, every request to the Braintree servers verifies the SSL connection
-    using the `PycURL <http://pycurl.sourceforge.net/>`_
-    library.  This ensures valid encryption of data and prevents man-in-the-middle attacks.
-
-    If you are in an environment where you absolutely cannot load `PycURL <http://pycurl.sourceforge.net/>`_, you
-    can turn off SSL Verification by setting::
-
-        Configuration.use_unsafe_ssl = True
-
-    This is highly discouraged, however, since it leaves you susceptible to
-    man-in-the-middle attacks.
-
-    If you are using Google App Engine, you must use unsafe ssl [1]_::
-
-        The proxy the URL Fetch service uses cannot authenticate the host it
-        is contacting. Because there is no certificate trust chain, the proxy
-        accepts all certificates, including self-signed certificates. The
-        proxy server cannot detect "man in the middle" attacks between App
-        Engine and the remote host when using HTTPS.
-
-.. [1] `URL Fetch Python API Overview <https://developers.google.com/appengine/docs/python/urlfetch/overview>`_
     """
-    use_unsafe_ssl = False
-
     @staticmethod
-    def configure(environment, merchant_id, public_key, private_key, http_strategy=None):
+    def configure(environment, merchant_id, public_key, private_key, **kwargs):
         Configuration.environment = environment
         Configuration.merchant_id = merchant_id
         Configuration.public_key = public_key
         Configuration.private_key = private_key
-        Configuration.default_http_strategy = http_strategy
+        Configuration.default_http_strategy = kwargs.get("http_strategy", None)
+        Configuration.timeout = kwargs.get("timeout", 60)
+        Configuration.wrap_http_exceptions = kwargs.get("wrap_http_exceptions", False)
 
     @staticmethod
-    def for_partner(environment, partner_id, public_key, private_key, http_strategy=None):
+    def for_partner(environment, partner_id, public_key, private_key, **kwargs):
         return Configuration(
             environment=environment,
             merchant_id=partner_id,
             public_key=public_key,
             private_key=private_key,
-            http_strategy=http_strategy
+            http_strategy=kwargs.get("http_strategy", None),
+            timeout=kwargs.get("timeout", 60),
+            wrap_http_exceptions=kwargs.get("wrap_http_exceptions", False)
         )
 
     @staticmethod
@@ -70,23 +47,29 @@ class Configuration(object):
             merchant_id=Configuration.merchant_id,
             public_key=Configuration.public_key,
             private_key=Configuration.private_key,
-            http_strategy=Configuration.default_http_strategy
+            http_strategy=Configuration.default_http_strategy,
+            timeout=Configuration.timeout,
+            wrap_http_exceptions=Configuration.wrap_http_exceptions
         )
 
     @staticmethod
     def api_version():
         return "3"
 
-    def __init__(self, environment, merchant_id, public_key, private_key, http_strategy=None):
+    def __init__(self, environment, merchant_id, public_key, private_key, **kwargs):
         self.environment = environment
         self.merchant_id = merchant_id
         self.public_key = public_key
         self.private_key = private_key
+        self.timeout = kwargs.get("timeout", 60)
+        self.wrap_http_exceptions = kwargs.get("wrap_http_exceptions", False)
+
+        http_strategy = kwargs.get("http_strategy", None)
 
         if http_strategy:
             self._http_strategy = http_strategy(self, self.environment)
         else:
-            self._http_strategy = self.__determine_http_strategy()
+            self._http_strategy = self.http()
 
     def base_merchant_path(self):
         return "/merchants/" + self.merchant_id
@@ -98,27 +81,4 @@ class Configuration(object):
         return braintree.util.http.Http(self)
 
     def http_strategy(self):
-        if Configuration.use_unsafe_ssl:
-            return braintree.util.http_strategy.httplib_strategy.HttplibStrategy(self, self.environment)
-        else:
-            return self._http_strategy
-
-    def __determine_http_strategy(self):
-        if "PYTHON_HTTP_STRATEGY" in os.environ:
-            return self.__http_strategy_from_environment()
-
-        if sys.version_info[0] == 2 and sys.version_info[1] == 5:
-            return braintree.util.http_strategy.pycurl_strategy.PycurlStrategy(self, self.environment)
-        else:
-            return braintree.util.http_strategy.requests_strategy.RequestsStrategy(self, self.environment)
-
-    def __http_strategy_from_environment(self):
-        strategy_name = os.environ["PYTHON_HTTP_STRATEGY"]
-        if strategy_name == "httplib":
-            return braintree.util.http_strategy.httplib_strategy.HttplibStrategy(self, self.environment)
-        elif strategy_name == "pycurl":
-            return braintree.util.http_strategy.pycurl_strategy.PycurlStrategy(self, self.environment)
-        elif strategy_name == "requests":
-            return braintree.util.http_strategy.requests_strategy.RequestsStrategy(self, self.environment)
-        else:
-            raise ValueError("invalid http strategy")
+        return self._http_strategy