+++ tests/unit/test_configuration.py
@@ -1,12 +1,13 @@
 from tests.test_helper import *
 import braintree
 import os
+import imp
 
 class TestConfiguration(unittest.TestCase):
     def test_works_with_unconfigured_configuration(self):
         try:
             # reset class level attributes on Configuration set in test helper
-            reload(braintree.configuration)
+            imp.reload(braintree.configuration)
             config = Configuration(
                 environment=braintree.Environment.Sandbox,
                 merchant_id='my_merchant_id',
@@ -14,13 +15,13 @@ class TestConfiguration(unittest.TestCase):
                 private_key='private_key'
             )
             config.http_strategy()
-        except AttributeError, e:
-            print e
+        except AttributeError as e:
+            print(e)
             self.assertTrue(False)
         finally:
             # repopulate class level attributes on Configuration
             import tests.test_helper
-            reload(tests.test_helper)
+            imp.reload(tests.test_helper)
 
     def test_base_merchant_path_for_development(self):
         self.assertEqual("/merchants/integration_merchant_id", Configuration.instantiate().base_merchant_path())
@@ -47,46 +48,16 @@ class TestConfiguration(unittest.TestCase):
         self.assertEqual(config.public_key, 'public_key')
         self.assertEqual(config.private_key, 'private_key')
 
-    def test_overriding_http_strategy_blows_up_if_setting_an_invalid_strategy(self):
-        old_http_strategy = None
-
-        if "PYTHON_HTTP_STRATEGY" in os.environ:
-            old_http_strategy = os.environ["PYTHON_HTTP_STRATEGY"]
-
-        try:
-            os.environ["PYTHON_HTTP_STRATEGY"] = "invalid"
-            strategy = Configuration.instantiate().http_strategy()
-            self.assertTrue(False, "Expected StandardError")
-        except ValueError, e:
-            self.assertEqual("invalid http strategy", e.message)
-        finally:
-            if old_http_strategy == None:
-                del(os.environ["PYTHON_HTTP_STRATEGY"])
-            else:
-                os.environ["PYTHON_HTTP_STRATEGY"] = old_http_strategy
-
-    def test_overriding_http_strategy(self):
-        old_http_strategy = None
-
-        if "PYTHON_HTTP_STRATEGY" in os.environ:
-            old_http_strategy = os.environ["PYTHON_HTTP_STRATEGY"]
-
-        try:
-            os.environ["PYTHON_HTTP_STRATEGY"] = "httplib"
-            strategy = Configuration.instantiate().http_strategy()
-            self.assertTrue(isinstance(strategy, braintree.util.http_strategy.httplib_strategy.HttplibStrategy))
-        finally:
-            if old_http_strategy == None:
-                del(os.environ["PYTHON_HTTP_STRATEGY"])
-            else:
-                os.environ["PYTHON_HTTP_STRATEGY"] = old_http_strategy
-
     def test_configuring_with_an_http_strategy(self):
         old_http_strategy = Configuration.default_http_strategy
 
+        class FakeStrategy(object):
+            def __init__(self, config, environment):
+                pass
+
         try:
-            Configuration.default_http_strategy = braintree.util.http_strategy.httplib_strategy.HttplibStrategy
+            Configuration.default_http_strategy = FakeStrategy
             strategy = Configuration.instantiate().http_strategy()
-            self.assertTrue(isinstance(strategy, braintree.util.http_strategy.httplib_strategy.HttplibStrategy))
+            self.assertTrue(isinstance(strategy, FakeStrategy))
         finally:
             Configuration.default_http_strategy = old_http_strategy